test: split the 11 suites over the 350-line ceiling; enforce it
The 350-line hard ceiling was applied to test files at WARNING
(#558) because 11 suites were already over it. Split each along
cohesive suite/struct boundaries into focused sibling files, moving
every test body verbatim (test-count parity checked per suite) and
copying per-file private helpers per the .claude/rules/tests.md
convention -- no new shared harness. Then promote the test-file
case in scripts/lint.sh from WARNING to ERROR, so a test file now
fails identically to a source file at the ceiling.

Splits (original -> resulting files, all <=350):
- MonocleTests 972 -> MonocleTests 208 + MonocleCommand/Cycle/
  Grouping + AppBarSlotSizing/AppBarScrollLayout (the last two
  were App-Bar math mis-homed in the Monocle file)
- DragTests 781 -> DragTests 206 + DragDrop + DragTargetAndVisual
- CommandTests 597 -> CommandTests 178 + CommandBehavior/Config/
  Navigation
- StateTests 485 -> StateTests 309 + StateManagers
- MoveToSpaceTests 434 -> MoveToSpace 179 + MoveToSpaceAnimation
  Gate + MoveToSpaceDesktopYield
- LayoutMiscTests 388 -> LayoutMisc 119 + FloatRuleRetile
- StickyCommandTests 385 -> StickyCommand 211 + StickyFocus
  Persistence
- TrackCommandsTests 377 -> TrackCommands 226 + TrackModeIntegration
- LayoutTests 366 -> LayoutBspTests + LayoutStackTests (emptied,
  removed)
- TrackLayoutTests 354 -> TrackLayout 256 + TrackPartition
- ProfileTests 351 -> ProfileTests 200 + ProfileSpacesPersistence

The soft 250 target stays source-only. Full suite green (2088
tests), build and lint clean.

Fixes #560
