+++ Tests/KiwiDeskCoreTests/DragTests.swift
@@ -204,578 +204,3 @@ struct DragCoordinatorTests {
         #expect(ended == [frame])
     }
 }
-
-@Suite("Drag drop resolution", .serialized)
-@MainActor
-struct DragDropTests {
-    private func makeCore() -> KiwiCore {
-        makeTestCore(
-            configDirectory: FileManager.default
-                .temporaryDirectory
-                .appendingPathComponent(
-                    "kiwi-drag-\(UUID().uuidString)"
-                )
-        )
-    }
-
-    private func addWindow(
-        _ core: KiwiCore,
-        _ raw: UInt32
-    ) {
-        core.state.apply(
-            .windowCreated(
-                ManagedWindow(
-                    id: WindowID(raw),
-                    pid: 1,
-                    appName: "App\(raw)"
-                )
-            )
-        )
-    }
-
-    @Test("Dropping onto another slot swaps the windows")
-    func swapOnDrop() throws {
-        // Needs a real screen for slot geometry.
-        guard NSScreen.main != nil else { return }
-        let core = makeCore()
-        addWindow(core, 1)
-        addWindow(core, 2)
-        let slots = core.tiler.calculatedFrames(
-            state: core.state
-        )
-        let home = try #require(slots[WindowID(1)])
-        let targetSlot = try #require(slots[WindowID(2)])
-
-        // Same size as at gesture start (a move), centered
-        // over the other window's slot.
-        var frame = home
-        frame.origin = CGPoint(
-            x: targetSlot.midX - home.width / 2,
-            y: targetSlot.midY - home.height / 2
-        )
-        // The drop target follows the cursor, not the frame
-        // center (#492): place the pointer over the target slot.
-        // `cursorLocation` is Cocoa; the flip is its own inverse,
-        // so `axPoint` of the AX center yields the Cocoa point
-        // that flips back to it.
-        core.drag.cursorLocation = {
-            GeometryUtils.axPoint(
-                CGPoint(x: targetSlot.midX, y: targetSlot.midY)
-            )
-        }
-        core.handleDragEnd(
-            WindowID(1),
-            start: home,
-            frame: frame
-        )
-        #expect(
-            core.state.workspaces[SpaceID(1)]?.windows
-                == [WindowID(2), WindowID(1)]
-        )
-        // The dropped window ends up focused.
-        #expect(core.activeSpace?.focused == WindowID(1))
-    }
-
-    @Test("Dropping into nowhere keeps the order (snap back)")
-    func snapBack() throws {
-        guard NSScreen.main != nil else { return }
-        let core = makeCore()
-        addWindow(core, 1)
-        addWindow(core, 2)
-
-        let nowhere = CGRect(
-            x: -9000,
-            y: -9000,
-            width: 100,
-            height: 100
-        )
-        // Cursor over no slot either — the drop target is keyed
-        // on the pointer (#492), so pin it away from every slot.
-        core.drag.cursorLocation = {
-            GeometryUtils.axPoint(CGPoint(x: -9000, y: -9000))
-        }
-        core.handleDragEnd(
-            WindowID(1),
-            start: nowhere.offsetBy(dx: 500, dy: 500),
-            frame: nowhere
-        )
-        #expect(
-            core.state.workspaces[SpaceID(1)]?.windows
-                == [WindowID(1), WindowID(2)]
-        )
-    }
-
-    @Test("Live preview shows ghost and drop zone over a slot")
-    func livePreview() throws {
-        guard NSScreen.main != nil else { return }
-        let core = makeCore()
-        addWindow(core, 1)
-        addWindow(core, 2)
-        let slots = core.tiler.calculatedFrames(
-            state: core.state
-        )
-        let home = try #require(slots[WindowID(1)])
-        let targetSlot = try #require(slots[WindowID(2)])
-
-        // Same size (a move, not a resize), centered over the
-        // other window's slot.
-        var frame = home
-        frame.origin = CGPoint(
-            x: targetSlot.midX - home.width / 2,
-            y: targetSlot.midY - home.height / 2
-        )
-        // Cursor over the target slot: the drop zone is keyed on
-        // the pointer, not the frame center (#492).
-        core.drag.cursorLocation = {
-            GeometryUtils.axPoint(
-                CGPoint(x: targetSlot.midX, y: targetSlot.midY)
-            )
-        }
-        core.handleDragMove(
-            WindowID(1),
-            start: home,
-            frame: frame
-        )
-        #expect(core.dragOverlay.isGhostVisible)
-        #expect(core.dragOverlay.isDropZoneVisible)
-
-        // Cursor over no slot: the drop zone goes away, the ghost
-        // (home slot) stays.
-        core.drag.cursorLocation = {
-            GeometryUtils.axPoint(CGPoint(x: -9000, y: -9000))
-        }
-        core.handleDragMove(
-            WindowID(1),
-            start: home,
-            frame: frame.offsetBy(dx: -9000, dy: -9000)
-        )
-        #expect(core.dragOverlay.isGhostVisible)
-        #expect(!core.dragOverlay.isDropZoneVisible)
-
-        // The drop hides everything.
-        core.handleDragEnd(
-            WindowID(1),
-            start: home,
-            frame: frame
-        )
-        #expect(!core.dragOverlay.isGhostVisible)
-        #expect(!core.dragOverlay.isDropZoneVisible)
-    }
-
-    @Test(
-        """
-        The gesture's first frame (no motion) shows no preview \
-        so a resize can't flash the ghost (#237)
-        """
-    )
-    func firstFrameShowsNoPreview() throws {
-        guard NSScreen.main != nil else { return }
-        let core = makeCore()
-        addWindow(core, 1)
-        addWindow(core, 2)
-        let slots = core.tiler.calculatedFrames(
-            state: core.state
-        )
-        let home = try #require(slots[WindowID(1)])
-        // start == frame: nothing has translated or resized yet,
-        // so the gesture reads as neither a move nor a resize.
-        // A resize begins here — showing the ghost now would be
-        // the one-frame flash.
-        core.handleDragMove(
-            WindowID(1),
-            start: home,
-            frame: home
-        )
-        #expect(!core.dragOverlay.isGhostVisible)
-        #expect(!core.dragOverlay.isDropZoneVisible)
-    }
-
-    @Test(
-        """
-        An edge-grabbed sub-threshold resize hides the ghost, \
-        while the same drag grabbed in the body shows it (#237)
-        """
-    )
-    func edgePressDiscriminatesSmallResize() throws {
-        guard NSScreen.main != nil else { return }
-        let core = makeCore()
-        addWindow(core, 1)
-        addWindow(core, 2)
-        let slots = core.tiler.calculatedFrames(
-            state: core.state
-        )
-        let home = try #require(slots[WindowID(1)])
-        // A < 10 pt width growth, below the magnitude
-        // threshold, with the origin nudged so movedOrigin is
-        // true — this isolates the pressNearEdge && sizeChanged
-        // path from the first-frame guard.
-        var frame = home
-        frame.size.width += 4
-        frame.origin.x += 1
-
-        // Press on the right edge — a resize grab: hidden.
-        core.mouse.seedPress(
-            at: CGPoint(x: home.maxX, y: home.midY)
-        )
-        core.handleDragMove(
-            WindowID(1),
-            start: home,
-            frame: frame
-        )
-        #expect(!core.dragOverlay.isGhostVisible)
-
-        // Same frames, but the press began in the body: the
-        // sub-threshold size wobble reads as move noise, so the
-        // ghost stays (the character-grid-move fallback).
-        core.mouse.seedPress(
-            at: CGPoint(x: home.midX, y: home.midY)
-        )
-        core.handleDragMove(
-            WindowID(1),
-            start: home,
-            frame: frame
-        )
-        #expect(core.dragOverlay.isGhostVisible)
-    }
-
-    @Test("Disabled toggles suppress the drag visuals")
-    func previewDisabled() throws {
-        guard NSScreen.main != nil else { return }
-        let core = makeCore()
-        core.tiler.settings.dragGhost.enabled = false
-        core.tiler.settings.dragDropZone.enabled = false
-        addWindow(core, 1)
-        addWindow(core, 2)
-        let slots = core.tiler.calculatedFrames(
-            state: core.state
-        )
-        let targetSlot = try #require(slots[WindowID(2)])
-
-        core.handleDragMove(
-            WindowID(1),
-            start: targetSlot,
-            frame: targetSlot
-        )
-        #expect(!core.dragOverlay.isGhostVisible)
-        #expect(!core.dragOverlay.isDropZoneVisible)
-    }
-
-    @Test("A resize-shaped drag shows no swap preview")
-    func previewSkipsResize() throws {
-        guard NSScreen.main != nil else { return }
-        let core = makeCore()
-        addWindow(core, 1)
-        addWindow(core, 2)
-        let slots = core.tiler.calculatedFrames(
-            state: core.state
-        )
-        let start = try #require(slots[WindowID(1)])
-        var frame = start
-        frame.size.width += 120
-
-        core.handleDragMove(
-            WindowID(1),
-            start: start,
-            frame: frame
-        )
-        #expect(!core.dragOverlay.isGhostVisible)
-        #expect(!core.dragOverlay.isDropZoneVisible)
-    }
-
-    @Test(
-        """
-        A moved window that never matched its slot's size \
-        still swaps (no false resize)
-        """
-    )
-    func constrainedWindowMoveIsNotAResize() throws {
-        guard NSScreen.main != nil else { return }
-        let core = makeCore()
-        addWindow(core, 1)
-        addWindow(core, 2)
-        let slots = core.tiler.calculatedFrames(
-            state: core.state
-        )
-        let home = try #require(slots[WindowID(1)])
-        let targetSlot = try #require(slots[WindowID(2)])
-
-        // The app refused the slot size (min size): the real
-        // frame is bigger than the slot for the whole
-        // gesture, but its size never CHANGES — a move.
-        var start = home
-        start.size.height += 80
-        var frame = start
-        frame.origin = CGPoint(
-            x: targetSlot.midX - frame.width / 2,
-            y: targetSlot.midY - frame.height / 2
-        )
-        // Pointer over the target slot: the swap is keyed on the
-        // cursor, not the (constrained) frame's center (#492).
-        core.drag.cursorLocation = {
-            GeometryUtils.axPoint(
-                CGPoint(x: targetSlot.midX, y: targetSlot.midY)
-            )
-        }
-        core.handleDragEnd(
-            WindowID(1),
-            start: start,
-            frame: frame
-        )
-        #expect(
-            core.state.workspaces[SpaceID(1)]?.windows
-                == [WindowID(2), WindowID(1)]
-        )
-    }
-}
-
-@Suite("Drag visual configuration", .serialized)
-@MainActor
-struct DragVisualCommandTests {
-    private func makeCore() -> KiwiCore {
-        makeTestCore(
-            configDirectory: FileManager.default
-                .temporaryDirectory
-                .appendingPathComponent(
-                    "kiwi-dragviz-\(UUID().uuidString)"
-                )
-        )
-    }
-
-    @Test("drag.* commands update every visual setting")
-    func commands() {
-        let core = makeCore()
-        core.execute(
-            "drag.set_ghost_enabled",
-            args: [.bool(false)]
-        )
-        core.execute(
-            "drag.set_ghost_fill_color",
-            args: [.string("#112233")]
-        )
-        core.execute(
-            "drag.set_ghost_border_width",
-            args: [.number(4)]
-        )
-        core.execute(
-            "drag.set_ghost_border_alignment",
-            args: [.string("outside")]
-        )
-        core.execute(
-            "drag.set_drop_zone_border_width",
-            args: [.number(3)]
-        )
-        core.execute(
-            "drag.set_drop_zone_border",
-            args: [.bool(false)]
-        )
-        core.execute(
-            "drag.set_corner_radius",
-            args: [.number(26)]
-        )
-        let settings = core.tiler.settings
-        #expect(!settings.dragGhost.enabled)
-        #expect(settings.dragGhost.fillColor == "#112233")
-        #expect(settings.dragGhost.borderWidth == 4)
-        #expect(settings.dragGhost.borderAlignment == .outside)
-        #expect(settings.dragDropZone.borderWidth == 3)
-        #expect(settings.dragDropZone.borderAlignment == .inside)
-        #expect(!settings.dragDropZone.border)
-        #expect(settings.dragCornerRadius == 26)
-    }
-
-    @Test("Bad colors and unknown settings are rejected")
-    func rejectsBadInput() {
-        let core = makeCore()
-        #expect(
-            !core.execute(
-                "drag.set_ghost_border_color",
-                args: [.string("kiwi")]
-            ).isSuccess
-        )
-        #expect(
-            !core.execute(
-                "drag.set_ghost_opacity",
-                args: [.number(1)]
-            ).isSuccess
-        )
-        #expect(
-            core.tiler.settings.dragGhost
-                == .ghostDefault
-        )
-    }
-}
-
-@Suite("Hex color parsing")
-struct HexColorTests {
-    @Test("6- and 8-digit hex parse, junk does not")
-    func parse() throws {
-        let rgb = try #require(
-            DragVisual.parseHex("#4E9F3D")
-        )
-        #expect(rgb.alpha == 1)
-        #expect(abs(rgb.green - 0x9F / 255.0) < 0.001)
-        let rgba = try #require(
-            DragVisual.parseHex("8B5E3C40")
-        )
-        #expect(abs(rgba.alpha - 0x40 / 255.0) < 0.001)
-        #expect(DragVisual.parseHex("#12345") == nil)
-        #expect(DragVisual.parseHex("kiwi") == nil)
-        #expect(DragVisual.parseHex("#GG5E3C") == nil)
-    }
-}
-
-@Suite("Drag target rule")
-struct DragTargetTests {
-    private let slots: [WindowID: CGRect] = [
-        WindowID(1): CGRect(x: 0, y: 0, width: 100, height: 100),
-        WindowID(2): CGRect(
-            x: 100,
-            y: 0,
-            width: 100,
-            height: 100
-        ),
-    ]
-
-    @Test("The slot under the cursor is the target")
-    func cursorRule() {
-        // Cursor over slot 2's area.
-        let cursor = CGPoint(x: 130, y: 50)
-        #expect(
-            DragTarget.swapTarget(
-                of: WindowID(1),
-                at: cursor,
-                slots: slots
-            ) == WindowID(2)
-        )
-    }
-
-    @Test(
-        """
-        Target follows the cursor onto the destination display \
-        even while the dragged window's center trails behind \
-        on the origin display (#492)
-        """
-    )
-    func cursorLeadsOverCenter() {
-        // A big window on the origin display dragged toward a
-        // smaller one. Its center still sits over the origin
-        // slot, but the cursor has already crossed to the
-        // destination slot.
-        let origin = CGRect(x: 0, y: 0, width: 2000, height: 1000)
-        let dest = CGRect(
-            x: 2000,
-            y: 0,
-            width: 800,
-            height: 600
-        )
-        let pool: [WindowID: CGRect] = [
-            WindowID(2): origin,
-            WindowID(3): dest,
-        ]
-        let windowCenter = CGPoint(x: 900, y: 500)
-        let cursor = CGPoint(x: 2400, y: 300)
-        // The old center rule would have picked the origin slot,
-        // never resolving the destination.
-        #expect(origin.contains(windowCenter))
-        #expect(!dest.contains(windowCenter))
-        // The cursor rule lands on the destination slot.
-        #expect(
-            DragTarget.swapTarget(
-                of: WindowID(1),
-                at: cursor,
-                slots: pool
-            ) == WindowID(3)
-        )
-    }
-
-    @Test("A window's own slot is never its target")
-    func neverSelf() {
-        // Cursor over slot 1 (the dragged window's own slot).
-        let cursor = CGPoint(x: 30, y: 30)
-        #expect(
-            DragTarget.swapTarget(
-                of: WindowID(1),
-                at: cursor,
-                slots: slots
-            ) == nil
-        )
-    }
-
-    @Test("A cursor over no slot yields no target")
-    func nowhere() {
-        let cursor = CGPoint(x: 500, y: 500)
-        #expect(
-            DragTarget.swapTarget(
-                of: WindowID(1),
-                at: cursor,
-                slots: slots
-            ) == nil
-        )
-    }
-
-    @Test(
-        """
-        Overlapping cascade slots resolve to the visible \
-        strip's owner, not dictionary order
-        """
-    )
-    func cascadeOverlap() {
-        // A stack overflow cascade: same column, each slot 40
-        // lower than the previous, all 100 tall — so slot n's
-        // visible part is the 40-point strip under its top.
-        let cascade: [WindowID: CGRect] = [
-            WindowID(2): CGRect(
-                x: 0,
-                y: 0,
-                width: 100,
-                height: 100
-            ),
-            WindowID(3): CGRect(
-                x: 0,
-                y: 40,
-                width: 100,
-                height: 100
-            ),
-            WindowID(4): CGRect(
-                x: 0,
-                y: 80,
-                width: 100,
-                height: 100
-            ),
-        ]
-        func target(centerY: CGFloat) -> WindowID? {
-            DragTarget.swapTarget(
-                of: WindowID(1),
-                at: CGPoint(x: 50, y: centerY),
-                slots: cascade
-            )
-        }
-        // Inside the top window's exposed strip.
-        #expect(target(centerY: 20) == WindowID(2))
-        // Two slots contain this point; the middle window's
-        // strip is the visible one.
-        #expect(target(centerY: 60) == WindowID(3))
-        // All three contain this point; the last window is
-        // raised on top of the whole cascade.
-        #expect(target(centerY: 90) == WindowID(4))
-    }
-
-    @Test("Identical slots pick a deterministic target")
-    func identicalSlots() {
-        let rect = CGRect(x: 0, y: 0, width: 100, height: 100)
-        let same: [WindowID: CGRect] = [
-            WindowID(2): rect,
-            WindowID(3): rect,
-        ]
-        let cursor = CGPoint(x: 50, y: 50)
-        for _ in 0..<10 {
-            #expect(
-                DragTarget.swapTarget(
-                    of: WindowID(1),
-                    at: cursor,
-                    slots: same
-                ) == WindowID(3)
-            )
-        }
-    }
-}
